requirements.txt
astroid==2.4.2
cycler==0.10.0
decorator==4.4.2
Fuzzy==1.2.2
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-fuzzy==0.4.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
toml==0.10.2
wrapt==1.12.1
